api/requirements.txt
fastapi==0.78.0
redis==4.3.3
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.19.4
uvicorn==0.17.6